adk/pyproject.toml
[project]
name = "cofacts-ai-agent"
version = "0.1.0"
description = "Cofacts AI multi-agent fact-checking system"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "pydantic",
    "google-adk",
    "google-genai",
    "httpx",
    "langfuse",
    "opentelemetry-sdk",
    "openinference-instrumentation-google-adk",
]
